feat: Add Tests for Vaccine Distribution Ledger Integrity
