+++ soroban/contracts/vaccine-distribution-ledger/src/lib.rs
@@ -9,7 +9,7 @@ mod distribution;
 mod utils;
 
 #[cfg(test)]
-mod test;
+mod tests;
 
 use soroban_sdk::{
     contract, contractimpl, Address, Env, String, Vec,
