+++ soroban/contracts/vaccine-distribution-ledger/src/tests/mod.rs
@@ -1,0 +1,4 @@
+pub mod ledger;
+pub mod verification;
+pub mod history;
+pub mod utils;
